Move fsdb; shipname to rl_loop/ subdirectory
