+++ oneoffs/heatmap.py
@@ -16,7 +16,7 @@ from absl import app, flags
 from tqdm import tqdm
 
 import go
-import fsdb
+from rl_loop import fsdb
 import oneoff_utils
